+++ OTAcceleratorCore/UI/OTOneToOneCommunicationController.h
@@ -1,0 +1,14 @@
+//
+//  MainViewController.h
+//
+// Copyright © 2016 Tokbox, Inc. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "OTAcceleratorSession.h"
+
+@interface OTOneToOneCommunicationController : UIViewController
+
++ (instancetype)oneToOneCommunicationControllerWithSession:(OTAcceleratorSession *)session;
+
+@end