+++ OTAcceleratorCoreSample/MainViewController.h
@@ -1,0 +1,13 @@
+//
+//  MainViewController.h
+//  OTAcceleratorCore
+//
+//  Created by Xi Huang on 4/7/17.
+//  Copyright © 2017 Tokbox, Inc. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface MainViewController : UITableViewController
+
+@end