+++ CHANGELOG.md
@@ -4,6 +4,16 @@ The changelog for `Accelerator-Core` iOS.
 
 --------------------------------------
 
+1.1.0
+-----
+
+### Enhancements
+
+- Add a Ready-in-Use for one-to-one communication: `OTOneToOneCommunicationController` 
+- Add two callback methods `didTapAudioButtonOnVideoControlView:` and `didTapVideoButtonOnVideoControlView:` on OTAudioVideoControlView
+- Seperate OTAudioVideoControlView from OTVideoView to be an independent class
+- Add a basic UI testing
+
 1.0.5
 -----